+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 * capture invalid option error, and display message
 * capture interrupt and clean up git commits
 
+### Bug fixes
+
+* when the version doesn't change (but dependencies do), do not attempt any more recursive updates
+
 ### New features
 
 ## 0.1.1 (2018-08-11)